@@ -1,34 +1,38 @@
 {
     "name": "ui-ng",
     "version": "0.0.0",
-    "license": "MIT",
     "scripts": {
+        "build": "ng build",
         "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build --prod",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
+        "postinstall": "ngcc",
+        "start": "ng serve --proxy-config proxy.json",
+        "start:insecure": "ng serve --proxy-config proxy.json --host 0.0.0.0 --disable-host-check"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^5.2.0",
-        "@angular/common": "^5.2.0",
-        "@angular/compiler": "^5.2.0",
-        "@angular/core": "^5.2.0",
-        "@angular/forms": "^5.2.0",
-        "@angular/http": "^5.2.0",
-        "@angular/platform-browser": "^5.2.0",
-        "@angular/platform-browser-dynamic": "^5.2.0",
-        "@angular/router": "^5.2.0",
-        "core-js": "^2.4.1",
-        "rxjs": "^5.5.6",
-        "zone.js": "^0.8.19"
+        "@angular/animations": "~11.2.4",
+        "@angular/common": "~11.2.4",
+        "@angular/compiler": "~11.2.4",
+        "@angular/core": "~11.2.4",
+        "@angular/forms": "~11.2.4",
+        "@angular/platform-browser": "~11.2.4",
+        "@angular/platform-browser-dynamic": "~11.2.4",
+        "@angular/router": "~11.2.4",
+        "@fortawesome/angular-fontawesome": "^0.8.2",
+        "@fortawesome/fontawesome-svg-core": "^1.2.35",
+        "@fortawesome/free-solid-svg-icons": "^5.15.3",
+        "bootstrap": "^4.6.0",
+        "bootswatch": "^4.6.0",
+        "jquery": "^3.6.0",
+        "rxjs": "~6.6.0",
+        "tslib": "^2.0.0",
+        "zone.js": "~0.11.3"
     },
     "devDependencies": {
-        "@angular/cli": "~1.7.3",
-        "@angular/compiler-cli": "^5.2.0",
-        "@angular/language-service": "^5.2.0",
-        "typescript": "~2.5.3"
+        "@angular-devkit/build-angular": "~0.1102.3",
+        "@angular/cli": "~11.2.3",
+        "@angular/compiler-cli": "~11.2.4",
+        "@types/node": "^12.11.1",
+        "typescript": "~4.1.5"
     }
 }
