@@ -1,29 +1,21 @@
-import { Component, Input, OnChanges, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
-    selector: 'ui-home-mockroutes',
+    selector: 'app-home-mockroutes',
     templateUrl: './mock-routes.component.html',
-    styles: []
 })
-export class PageHomeMockRoutesComponent implements OnChanges, OnInit {
-    @Input('mockRoutes') public mockRoutes: any = null;
-    @Input('server') public server: string = '';
+export class MockRoutesComponent implements OnInit {
+    //
+    // Properties.
 
+    //
+    // Construction.
     constructor() {
     }
-
-    ngOnChanges() {
-        if (this.mockRoutes && Array.isArray(this.mockRoutes)) {
-            for (const conf of this.mockRoutes) {
-                conf.hasGuards = false;
-                if (conf.routes) {
-                    conf.routes.forEach((route: any) => {
-                        conf.hasGuards = conf.hasGuards || route.guardPath !== null || route.guardName !== null;
-                    });
-                }
-            }
-        }
+    //
+    // Public methods.
+    public ngOnInit(): void {
     }
-    ngOnInit() {
-    }
+    //
+    // Protected methods.
 }
