@@ -1,5 +1,3 @@
 export const environment = {
-    production: true,
-    serverUri: '',
-    serverPort: false
+    production: true
 };
